server/requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv

